Move OnlinePlayersEntity and update inheritance
