+++ src/game/scenes/main/main-menu-scene.ts
@@ -2,7 +2,7 @@ import { CloseableMessageEntity } from "../../entities/common/closeable-message-
 import { MenuOptionEntity } from "../../entities/common/menu-option-entity.js";
 import { TitleEntity } from "../../entities/common/title-entity.js";
 import { ServerMessageWindowEntity } from "../../entities/server-message-window-entity.js";
-import { OnlinePlayersEntity } from "../../entities/common/online-players-entity.js";
+import { OnlinePlayersEntity } from "../../entities/online-players-entity.js";
 import { APIService } from "../../services/network/api-service.js";
 import type { MessagesResponse } from "../../interfaces/responses/messages-response.js";
 import { BaseGameScene } from "../../../core/scenes/base-game-scene.js";
